Refactor task update and delete methods
- Update the `updateTask` method in the `TaskController` class to include the user ID in the `updateTask` service call.
- Update the `deleteTask` method in the `TaskController` class to include the user ID in the `deleteTask` service call.
- Add validation middleware for updating tasks in the `updateTaskValidator` array in the `tasks.validator.ts` file.
- Update the task route to use the `updateTaskValidator` middleware for the `PATCH /tasks/:id` route.
- Update the `updateTask` method in the `TaskService` class to check if the user is the owner of the task before updating it.
- Update the `deleteTask` method in the `TaskService` class to check if the user is the owner of the task before deleting it.

+++ src/services/tasks/task.service.ts
@@ -1,6 +1,7 @@
 import { autoInjectable } from 'tsyringe';
 
 import { TaskDao } from '../../DB/dao/task.dao';
+import HttpException from '../../exceptions/HttpException';
 import { ITask } from '../../interfaces';
 
 @autoInjectable()
@@ -23,12 +24,22 @@ class TaskService {
     return newTask;
   };
 
-  updateTask = async (id: string, task: ITask) => {
-    const updatedTask = await this.taskDao.updateOneById(id, task);
+  updateTask = async (id: string, payload: ITask, userId: string | undefined) => {
+    // check if the user is the owner of the task
+    const task = await this.taskDao.getOneById(id);
+
+    if (!task) throw new HttpException(404, 'Task not found');
+    // convert the id to string to compare it with the ownerId
+    if (task.ownerId !== userId?.toString()) throw new HttpException(403, 'You are not allowed to update this task');
+    const updatedTask = await this.taskDao.updateOneById(id, payload);
     return updatedTask;
   };
 
-  deleteTask = async (id: string) => {
+  deleteTask = async (id: string, userId: string | undefined) => {
+    const task = await this.taskDao.getOneById(id);
+    if (!task) throw new HttpException(404, 'Task not found');
+    if (task.ownerId !== userId?.toString()) throw new HttpException(403, 'You are not allowed to update this task');
+
     const deletedTask = await this.taskDao.deleteOneById(id);
     return deletedTask;
   };
